Danny Monday work
I couldn't get the enemy to shoot when the player is detected so instead the enemy shoots on a timer.

added enemy health.

made it so enemy bullets can damage the player (does different damage than when the player simply runs into the enemy)

the player can kill the enemies when there hp reaches 0. kicking does more damage than shooting.